Fix storage.js and separate image assets
