+++ src/screens/loading/Loading.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ActivityIndicator, StyleSheet, View, Image } from 'react-native';
-import Logo from '../../../assets/logo.png';
+import Logo from '../../../assets/images/logo.png';
 import theme from '../../constants/theme';
 
 export default function loadingScreen() {
